Fix F7: minors folded in (clippy, own_kill strictness, EOF logging, provenance)
- Clippy: wispd/src/main.rs `seq % 5 == 0` -> `seq.is_multiple_of(5)`;
  wisp-hud/src/text.rs removed the redundant `baseline as i32` cast
  (baseline is already i32). cargo clippy --workspace --all-targets is
  now warning-free.
- wispd/src/rules.rs own_kill: the spec rule is `You have slain (.+)!`,
  requiring at least one character between the prefix and the trailing
  `!`; the old rule accepted a bare "You have slain !". Added a test.
  Verified against the reference fixture
  (eqlog_Daggo_freeport.txt): zero occurrences of "You have slain !", and
  "You have slain " still matches exactly 3,722 times -- the verified
  kill count is unaffected.
- wisp-hud: SnapshotStream gained had_error(), set when next_snapshot
  returns None because of a read error rather than a clean EOF. main.rs
  now prints "daemon closed the connection" only on a clean EOF, instead
  of unconditionally after every loop exit.
- PROVENANCE.md: added dated 2026-09-08 entries for Task 9 (layer-shell,
  verified against KWin with the stub daemon by attach/present + a pool
  size measurement proving the release-event fix; real-game no-gamescope
  run pending JDS300) and Task 10 (plain-window, verified by xprop
  readback on KWin XWayland; click-through pending JDS300; and the
  decorated-frame defect this final-review wave found and fixed in F1,
  with its readback).

+++ crates/wisp-hud/src/text.rs
@@ -71,7 +71,7 @@ impl Renderer {
                         continue;
                     }
                     let x = pen_x as i32 + metrics.xmin + gx as i32;
-                    let y = baseline as i32 - metrics.height as i32 - metrics.ymin + gy as i32;
+                    let y = baseline - metrics.height as i32 - metrics.ymin + gy as i32;
                     if x < 0 || y < 0 || x >= width as i32 || y >= height as i32 {
                         continue;
                     }
